Ensure interrupt status is bounds-check-free

+++ esp-hal/src/interrupt/mod.rs
@@ -198,13 +198,13 @@ impl InterruptStatus {
     }
 
     /// Is the given interrupt bit set
-    pub fn is_set(&self, interrupt: u16) -> bool {
-        (self.status[interrupt as usize / 32] & (1 << (interrupt as u32 % 32))) != 0
+    pub fn is_set(&self, interrupt: u8) -> bool {
+        (self.status[interrupt as usize / 32] & (1 << (interrupt % 32))) != 0
     }
 
     /// Set the given interrupt status bit
-    pub fn set(&mut self, interrupt: u16) {
-        self.status[interrupt as usize / 32] |= 1 << (interrupt as u32 % 32);
+    pub fn set(&mut self, interrupt: u8) {
+        self.status[interrupt as usize / 32] |= 1 << (interrupt % 32);
     }
 
     /// Return an iterator over the set interrupt status bits
